+++ src/coloring/set_color_matrix.c
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 19:04:24 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/28 19:25:23 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/29 17:11:26 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,10 +39,10 @@ int	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 	int				collisions;
 	t_figure_point	closest_point;
 
-	c_matrix = malloc(sizeof(int *) * HORIZ_RESOLUTION);
+	c_matrix = ft_calloc(sizeof(int *), HORIZ_RESOLUTION + 1);
 	x = -1;
 	while (++x < HORIZ_RESOLUTION)
-		c_matrix[x] = malloc(sizeof(int) * RESOLUTION);
+		c_matrix[x] = ft_calloc(sizeof(int), RESOLUTION + 1);
 	y = 0;
 	while (y < RESOLUTION)
 	{
@@ -65,3 +65,16 @@ int	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 	return (c_matrix);
 }
 
+void	free_array_colors(int ***list)
+{
+	int	a;
+
+	a = 0;
+	if (*list)
+	{
+		while ((*list)[a])
+			free((*list)[a++]);
+		free((*list)[a]);
+		free(*list);
+	}
+}