+++ src/main.c
@@ -13,6 +13,8 @@ static void	doing_calculations(char **argv)
 	v_matrix = set_cam_vectors(set_corners(get_item_by_type(&items, CAMERA)));
 	c_matrix = set_color_matrix(items, v_matrix);
 	render_image_on_mlx(c_matrix);
+	free_array_coords(&v_matrix);
+	free_array_colors(&c_matrix);
 
 //	lst_rt_print(items);
 //	get_items_by_type(&items2, items, SPHERE);