+++ src/parser/setting_cam_vectors.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   setting_cam_vectors.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 15:01:28 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/20 22:56:14 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/28 22:39:13 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,11 +74,11 @@ t_coords	**set_cam_vectors(t_corners corn)
 	t_coords	**matrix;
 	int			ctr;
 
-	matrix = malloc(sizeof(t_coords *) * HORIZ_RESOLUTION);
+	matrix = ft_calloc(sizeof(t_coords *), HORIZ_RESOLUTION + 1);
 	ctr = 0;
 	while (ctr < HORIZ_RESOLUTION)
 	{
-		matrix[ctr] = malloc(sizeof(t_coords) * RESOLUTION);
+		matrix[ctr] = ft_calloc(sizeof(t_coords), RESOLUTION + 1);
 		ctr++;
 	}
 	matrix[0][0] = corn.tl;
@@ -88,4 +88,18 @@ t_coords	**set_cam_vectors(t_corners corn)
 	set_cam_margins(matrix);
 	set_cam_middle(matrix);
 	return (matrix);
+}
+
+void	free_array_coords(t_coords ***list)
+{
+	int	a;
+
+	a = 0;
+	if (*list)
+	{
+		while ((*list)[a])
+			free((*list)[a++]);
+		free((*list)[a]);
+		free(*list);
+	}
 }